Chains SNDFs to MISCF responses
Periodically we were seeing corruption in MISCF responses.  This corruption tended to include portions of the letters "SNDF" leading me to believe that the send buffer on the device was being corrupted.  In an attempt to avoid this corruption now instead of emitting SNDF requests at regular intervals a new request will be emitted whenever a response is received.