Set up a separate domain participant for patient-agnostic applications and began consolidating patient-specific numerics to a single NumericFxList.  See details.
For the life of me I could not get multiple readers of the same topic in different subscribers to reliably emit instance lifecycle events when changing partitions.  So I set up a separate participant to house all entities that are patient agnostic.  Also converted the "VitalModel" to use the NumericFxList.  But I found similar bad behaviour amongst numerics with multiple readers so I've started to work on moving the DataCollector and HL7Emitter to use the same NumericFxList used by the other components